Refactor for implementing broker e2e tests
