+++ msal/src/test/java/com/microsoft/identity/client/RoboTestCacheHelper.java
@@ -22,7 +22,7 @@
 //  THE SOFTWARE.
 package com.microsoft.identity.client;
 
-import com.microsoft.identity.client.e2e.utils.TestConstants;
+import com.microsoft.identity.internal.testutils.TestConstants;
 import com.microsoft.identity.common.exception.ClientException;
 import com.microsoft.identity.common.internal.authorities.AccountsInOneOrganization;
 import com.microsoft.identity.common.internal.authorities.Authority;
