+++ msal/src/test/java/com/microsoft/identity/client/e2e/tests/PublicClientApplicationAbstractTest.java
@@ -29,7 +29,7 @@ import androidx.test.core.app.ApplicationProvider;
 
 import com.microsoft.identity.client.IPublicClientApplication;
 import com.microsoft.identity.client.PublicClientApplication;
-import com.microsoft.identity.client.e2e.utils.RoboTestUtils;
+import com.microsoft.identity.internal.testutils.RoboTestUtils;
 import com.microsoft.identity.client.exception.MsalException;
 
 import org.junit.Before;
@@ -39,6 +39,8 @@ import java.io.File;
 import static org.junit.Assert.fail;
 
 public abstract class PublicClientApplicationAbstractTest implements IPublicClientApplicationTest {
+
+    protected final String SHARED_PREFERENCES_NAME = "com.microsoft.identity.client.account_credential_cache";
 
     protected Context mContext;
     protected Activity mActivity;
