+++ msal/src/test/java/com/microsoft/identity/client/e2e/tests/mocked/SingleAccountAcquireTokenMockTest.java
@@ -28,10 +28,10 @@ import androidx.annotation.Nullable;
 import com.microsoft.identity.client.IAccount;
 import com.microsoft.identity.client.IPublicClientApplication;
 import com.microsoft.identity.client.ISingleAccountPublicClientApplication;
-import com.microsoft.identity.client.e2e.utils.RoboTestUtils;
+import com.microsoft.identity.internal.testutils.RoboTestUtils;
 import com.microsoft.identity.client.exception.MsalException;
 
-import static com.microsoft.identity.client.e2e.utils.TestConstants.Configurations.SINGLE_ACCOUNT_MODE_MOCK_TEST_CONFIG_FILE_PATH;
+import static com.microsoft.identity.internal.testutils.TestConstants.Configurations.SINGLE_ACCOUNT_MODE_MOCK_TEST_CONFIG_FILE_PATH;
 import static org.junit.Assert.fail;
 
 public class SingleAccountAcquireTokenMockTest extends AcquireTokenMockTest {
